Added scores!
Every second survived is 1 point.
Every barrel picked up is 100 points.
Every enemy killed is 500 points.

+++ Assets/GameOverController.cs
@@ -16,7 +16,7 @@ public class GameOverController : MonoBehaviour {
 
         if (PlayerController.main && PlayerController.main.currentFuel <= 0) message.text = "No Fuel";
 
-        string text = "Score : " + PlayerController.score;
+        string text = "Score : " + (int)PlayerController.score;
         string[] chars = text.Split();
         score.text = string.Join(" ", chars);
 	}
